+++ sperl_allocator_util.h
@@ -2,8 +2,8 @@
 #define SPERL_ALLOCATOR_UTIL_H
 
 void SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure();
-void* SPerl_ALLOCATOR_UTIL_safe_malloc(size_t count, size_t size);
-void* SPerl_ALLOCATOR_UTIL_safe_malloc_zero(size_t count, size_t size);
-void* SPerl_ALLOCATOR_UTIL_safe_realloc(void* ptr, size_t count, size_t size);
+void* SPerl_ALLOCATOR_UTIL_safe_malloc(int64_t count, int64_t size);
+void* SPerl_ALLOCATOR_UTIL_safe_malloc_zero(int64_t count, int64_t size);
+void* SPerl_ALLOCATOR_UTIL_safe_realloc(void* ptr, int64_t count, int64_t size);
 
 #endif
